bessel function of the second kind from boost library

Leapfrog routine

Leapfrog routine

besselK_boost(x, v)

besselK(x, v)

matern(nu, a, rho, tau, D)

trenchDetcpp(c)

trenchInvcpp(v)

loglikeGPcpp(Y, Z, A, logcovDet, sigmak, nk, D, Y2)

likelihoodGPcpp(Xk, tau, h, nk, D, materncov = 0L, nu = 2)

gradientrhomatern(Y, drvrhomatern, nk, D, Z, A, sigmak)

gradientamatern(Y, amatern, nk, D, Z, A, sigmak)

gradientGPcppmatern(Xk, tau, h, nk, D, nu)

LeapfrogGPcppPC(Xk, lambda, tau, p, x, m, nk, D, L, delta, nu)

sampleGPmeanmaterncpp(Xk, tau, h, nk, D, nu)

makeComponent(X, BX, Y, BY, j)

sampleGPmeancpp(Xk, tau, h, nk, D)

normalisedData(Xknown, BX, Xunknown, BXun, hypers, nk, tau, D, j)

normalisedDatamatern(Xknown, BX, Xunknown, BXun, hypers, nk, tau, D, j, nu)

centeredDatamatern(Xknown, BX, Xunknown, BXun, hypers, nk, tau, D, K, nu)

componentloglike(centereddata, sigmak)

comploglike(centereddata, sigmak)

comploglikelist(centereddata, sigmak)

sampleDirichlet(numSamples, alpha)

sampleOutliercpp(allocoutlierprob)

sampleAlloccpp(allocprob)

centeredData(Xknown, BX, Xunknown, BXun, hypers, nk, tau, D, K)

mahaInt(X, mu, sigma, isChol = FALSE)

dmvtInt(X, mu, cholDec, log, df)

dmvtCpp(X_, mu_, sigma_, df_, log_, isChol_)

gradientGPcpp(Xk, tau, h, nk, D)

LeapfrogGPcpp(Xk, tau, p, x, m, nk, D, L, delta)

rcpp_pgdraw(b, c)

Arguments

x

position

v

argument of trench algorithm

nu

smoothness parameter of matern covariance

a

amplitude

rho

length-scale

tau

indexing term

D

number of samples

c

parameter of PG distribution

Y

pointer to data to be subset. X and Y will be joined

Z

special matrix from trench algorithm (see Crook et al arxiv 2019)

A

special matrix from trench algorithm (see Crook et al arxiv 2019)

logcovDet

log determine of the covariancematrix

sigmak

variance term

nk

number of observations

Y2

vectorised data (see Crook et al arxiv 2019)

Xk

The data

h

vector of hyperparamters

materncov

logical indicating whether to use matern or gaussian covariance. Defaults to Guassian covariance

drvrhomatern

deterivate of matern covariance wrt to rho

amatern

deterivate of matern covariance wrt to amplitude

lambda

parameters of penalised complexity prior

p

momentum

m

mass

L

iterations

delta

stepsize

X

data

BX

indexing set to make component

BY

pointer to subsetting matrix

j

indicator of localisations i.e. niche j

Xknown

data with known localisations

Xunknown

data with unknown localisations

BXun

indexing set for unknown localisations

hypers

vector of hyperparameters

K

number of components

centereddata

pointer to centered data

numSamples

The number of samples desired

alpha

The concentration parameter

allocoutlierprob

The probabilities of being allocated to the outlier component

allocprob

probability of being allocated to particular component

mu

mean

sigma

variance matrix

isChol

boolen indicated whether sigma is cholesky decomposition

cholDec

Cholesky decomposition of variance matrix

log

boolen of log density

df

degrees of freedom for t distribution

X_

the data

mu_

the mean

sigma_

the variance matrix

df_

the degrees of freedom

log_

return log density (boolean).

isChol_

is variance matrix in cholesky decomposition

b

parameter of PG distribution

Value

A numeric indicating the density of the t-distribution

Examples

dmvtCpp(diag(1,1,1), 1, diag(1,1,1), 1, TRUE, TRUE)
#> [1] -1.14473