Computes Organelle means and variances using markers

Computes Organelle means and variances using markers

meanOrganelle(object, fcol = "markers")

Arguments

object

a instance of class MSnset

fcol

a feature column indicating which feature define the markers

Value

returns a list of means and variances for each

Examples

library(pRolocdata)
data("tan2009r1")
meanOrganelle(object = tan2009r1)
#> $M
#>                     [,1]       [,2]       [,3]      [,4]
#> Cytoskeleton  0.32452971 0.13363686 0.16035714 0.3812560
#> ER            0.09245588 0.26202243 0.49546539 0.1500950
#> Golgi         0.09011285 0.45692300 0.31440777 0.1385308
#> Lysosome      0.17331250 0.17410412 0.30210413 0.3504584
#> Nucleus       0.09862194 0.08920622 0.13819010 0.6739631
#> PM            0.31488253 0.29246644 0.24172762 0.1509448
#> Peroxisome    0.08750000 0.05550000 0.30950000 0.5476250
#> Proteasome    0.39229667 0.16374667 0.15180667 0.2920100
#> Ribosome 40S  0.21583255 0.14110145 0.15042710 0.4925910
#> Ribosome 60S  0.15188522 0.20733506 0.27261641 0.3681730
#> mitochondrion 0.34107938 0.08752808 0.08649502 0.4848726
#> 
#> $V
#>                       [,1]         [,2]         [,3]         [,4]
#> Cytoskeleton  0.0005567348 0.0003577764 0.0001421012 0.0004165805
#> ER            0.0003569855 0.0037328532 0.0032790437 0.0005888035
#> Golgi         0.0004977598 0.0274890254 0.0178810272 0.0016822246
#> Lysosome      0.0003712098 0.0013112590 0.0009604177 0.0012664219
#> Nucleus       0.0015571281 0.0005456730 0.0008441521 0.0032123124
#> PM            0.0017997835 0.0022816101 0.0015578345 0.0010543192
#> Peroxisome    0.0009583333 0.0006650000 0.0005430000 0.0009505625
#> Proteasome    0.0015753773 0.0002980741 0.0003121035 0.0011100601
#> Ribosome 40S  0.0014765452 0.0006973985 0.0006128625 0.0018266481
#> Ribosome 60S  0.0014816608 0.0028586108 0.0015856400 0.0029579196
#> mitochondrion 0.0011069696 0.0005501043 0.0004908850 0.0008557990
#>