Infrastructure to to store and process MCMC results

The bandleParams infrastructure is used to store and process MCMC results for bandle model from Crook et al 2021

chains(object)

# S4 method for bandleParams
show(object)

# S4 method for nicheParam
show(object)

# S4 method for bandleChain
show(object)

# S4 method for bandleChains
length(x)

# S4 method for bandleParams
length(x)

# S4 method for bandleSummaries
length(x)

# S4 method for nicheParams
length(x)

# S4 method for nicheParams
length(x)

posteriorEstimates(object)

# S4 method for bandleSummary
posteriorEstimates(object)

summaries(object)

params(object)

bandleJoint(object)

# S4 method for bandleSummary
bandleJoint(object)

# S4 method for bandleChains,ANY,ANY
[[(x, i, j = "missing", drop = "missing")

# S4 method for bandleParams,ANY,ANY
[[(x, i, j = "missing", drop = "missing")

# S4 method for bandleChains,ANY,ANY,ANY
[(x, i, j = "missing", drop = "missing")

# S4 method for bandleParams,ANY,ANY,ANY
[(x, i, j = "missing", drop = "missing")

# S4 method for bandleChains
show(object)

# S4 method for bandleSummaries
show(object)

# S4 method for bandleSummaries,ANY,ANY
[[(x, i, j = "missing", drop = "missing")

# S4 method for bandleSummaries,ANY,ANY
[[(x, i, j = "missing", drop = "missing")

# S4 method for bandleSummaries,ANY,ANY,ANY
[(x, i, j = "missing", drop = "missing")

# S4 method for nicheParams,ANY,ANY
[[(x, i, j = "missing", drop = "missing")

# S4 method for nicheParams,ANY,ANY
[[(x, i, j = "missing", drop = "missing")

# S4 method for nicheParams,ANY,ANY,ANY
[(x, i, j = "missing", drop = "missing")

# S4 method for nicheParams
show(object)

Arguments

object

object of class nicheParams.

x

Object to be subset.

i

An integer(). Should be of length 1 for [[.

j

Missing.

drop

Missing.

Value

An object of class bandleParams which stores the main results for the analysis when using bandle

Details

Objects of the bandleParams class are created with the bandle() function These objects store the priors for the model and the results of the MCMC chains, which themselves are stored as an instance of class bandleChains and can be accessed with the chains() function. A summary of the bandleChains (or class bandleSummary) can be further computed with the bandleProcess function.

see the bandle vignette for examples

Slots

chains

list() containing the individual full MCMC chain results in an bandleChains instance. Each element must be a valid bandleChain instance.

posteriorEstimates

A DataFrame documenting the posteriors in an bandleSummary instance

diagnostics

A matrix of dimensions 1 by 2 containing the bandleSummary diagnostics.

bandle.joint

A matrix of dimensions N by K storing the joint probability in an bandleSummary instance for each of the first condition

chains

list() containing the individual bandle Summaries for different conditions results in an bandleSummaries instance. Each element must be a valid bandleSummary instance.

method

A character() storing the bandle method name

priors

A list() with the priors for the parameters

seed

An integer() with the random number generation seed.

summary

Object of class bandleSummary the summarised MCMC results available in the bandleParams instance.

chains

Object of class bandleChains containing the full MCMC results in the bandleParams instance

datset

character indicating which dataset i.e control or treatment

replicate

integer an integer indicating which replicate

K

integer(1) indicating the number of components.

D

integer(1) indicating the number of samples.

method

character(1) defining the method used. Currently bandle

mk

matrix(K, D)

lambdak

numeric(K)

nuk

numeric(K)

sk

array(K, D, D)

params

list() containing the individual nicheParam objects results in an bandleParams instance. Each element must be a valid bandleParam instance.

dataset

character indicating the dataset usaully control or treatment

replicate

integer indicating the number of dataset replicate

n

integer(1) indicating the number of MCMC interactions. Stored in an bandleChain instance.

K

integer(1) indicating the number of components. Stored in an bandleChain instance.

N

integer(1) indicating the number of proteins. Stored in an bandleChain instance.

niche

matrix(N, n) component allocation results of an bandleChain instance.

nicheProb

matrix(N, n, K) component allocation probabilities of an bandleChain instance.

outlier

matrix(N, n) outlier allocation results.

outlierProb

matrix(N, n, 2) outlier allocation probabilities of an bandleChain instance.